Removed allegro_utils.h dependency
